Create Benchmark Page File Validation
Create Benchmark Page File Validation. I have added checks to see if the file is in the appropriate format for the  Create Benchmark page. Also, I fixed a bug causing an extra column to be added to the benchmark file when it is created. This extra column was the row names.